Renaming sf_ransac to sf_xyz
- Wrap ransac scalar fields in a more common module for scalar fields
that only require xyz.
